+++ definitions/tuya/TS0013.cpp
@@ -17,6 +17,8 @@ constexpr Expose kAutoExposes[] = {
 
 constexpr BindingSpec kAutoBindings[] = {
     {1, 0x0006},
+    {2, 0x0006},
+    {3, 0x0006},
 };
 // --- end auto-generated block ---